[NO ISSUE][OTH] Cleanup Redundant Type Cast
- user model changes: no
- storage format changes: no
- interface changes: no

Reviewed-on: https://asterix-gerrit.ics.uci.edu/2887
Sonar-Qube: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Contrib: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ asterixdb/asterix-om/src/main/java/org/apache/asterix/om/base/AInterval.java
@@ -71,7 +71,7 @@ public class AInterval implements IAObject {
 
     @Override
     public int hashCode() {
-        return (int) (((int) (this.intervalStart ^ (this.intervalStart >>> 32))) * 31
+        return (((int) (this.intervalStart ^ (this.intervalStart >>> 32))) * 31
                 + (int) (this.intervalEnd ^ (this.intervalEnd >>> 32))) * 31 + (int) this.typetag;
     }
